[UPD] Admin page [ADD] Remove manager - Not Dry
